refactor: infra/ai/dto 페키지 세분화
