+++ hertz-be/src/main/java/com/hertz/hertz_be/global/batch/TuningReportGenerationProcessor.java
@@ -2,9 +2,8 @@ package com.hertz.hertz_be.global.batch;
 
 import com.hertz.hertz_be.domain.channel.entity.SignalRoom;
 import com.hertz.hertz_be.domain.channel.repository.SignalMessageRepository;
-import com.hertz.hertz_be.domain.channel.repository.SignalRoomRepository;
 import com.hertz.hertz_be.domain.user.repository.UserRepository;
-import com.hertz.hertz_be.global.infra.ai.dto.AiTuningReportGenerationRequest;
+import com.hertz.hertz_be.global.infra.ai.dto.request.AiTuningReportGenerationRequest;
 import com.hertz.hertz_be.global.infra.ai.support.UserDataAssembler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.item.ItemProcessor;
